# AccSol.React/Data/swagger2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Accsol.API",
    "version": "v1"
  },
  "paths": {
...
    "/JournalEntries/Get": {
      "post": {
        "tags": [
          "JournalEntries"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Journal"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Journal"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Journal"
                }
              }
            }
          }
        }
      }
    },
    "/JournalEntries/Save": {
      "post": {
        "tags": [
          "JournalEntries"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Journal"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Journal"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Journal"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/JournalEntries/Delete": {
      "post": {
        "tags": [
          "JournalEntries"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/PayeeCategories/Get": {
      "post": {
        "tags": [
          "PayeeCategories"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PayeeCategory"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayeeCategory"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayeeCategory"
                }
              }
            }
          }
        }
      }
    },
    "/PayeeCategories/Save": {
      "post": {
        "tags": [
          "PayeeCategories"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayeeCategory"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PayeeCategory"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PayeeCategory"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/PayeeCategories/Delete": {
      "post": {
        "tags": [
          "PayeeCategories"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/Payees/Get": {
      "post": {
        "tags": [
          "Payees"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Payee"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payee"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payee"
                }
              }
            }
          }
        }
      }
    },
    "/Payees/Save": {
      "post": {
        "tags": [
          "Payees"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Payee"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Payee"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Payee"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Payees/Delete": {
      "post": {
        "tags": [
          "Payees"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/PettyCashes/Get": {
      "post": {
        "tags": [
          "PettyCashes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PettyCash"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PettyCash"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PettyCash"
                }
              }
            }
          }
        }
      }
    },
    "/PettyCashes/Save": {
      "post": {
        "tags": [
          "PettyCashes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PettyCash"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PettyCash"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PettyCash"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PettyCash"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PettyCash"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PettyCash"
                }
              }
            }
          }
        }
      }
    },
    "/PettyCashes/Delete": {
      "post": {
        "tags": [
          "PettyCashes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/PettyCashes/DownloadExcelReport": {
      "post": {
        "tags": [
          "PettyCashes"
        ],
        "parameters": [
          {
            "name": "excelFilePath",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xlsUrl",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PettyCashSummaryReportVMGroupedReportData"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PettyCashSummaryReportVMGroupedReportData"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PettyCashSummaryReportVMGroupedReportData"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Journal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "pettyCashID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "coaID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "debit": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "credit": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Payee": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "payeeCategoryID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "employeeID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PayeeCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PettyCash": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "pcfNo": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "payeeID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "particulars": {
            "type": "string",
            "nullable": true
          },
          "projectCodeID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "departmentID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "clientID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PettyCashSummaryReportVM": {
        "type": "object",
        "properties": {
          "row": {
            "type": "integer",
            "format": "int32"
          },
          "pcfNo": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "payee": {
            "type": "string",
            "nullable": true
          },
          "departmentName": {
            "type": "string",
            "nullable": true
          },
          "particulars": {
            "type": "string",
            "nullable": true
          },
          "clientName": {
            "type": "string",
            "nullable": true
          },
          "projectCode": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "accountName": {
            "type": "string",
            "nullable": true
          },
          "debit": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "credit": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PettyCashSummaryReportVMGroupedReportData": {
        "type": "object",
        "properties": {
          "totals": {
            "type": "object",
            "additionalProperties": {},
            "nullable": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PettyCashSummaryReportVM"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}